+++ CHANGELOG.md
@@ -15,6 +15,8 @@ We refer to [GitHub issues](https://github.com/JabRef/jabref/issues) by using `#
 - [#1026](https://github.com/JabRef/jabref/issues/1026) JabRef does no longer delete user comments outside of BibTeX entries and strings
 
 ### Fixed
+- Fixed [#1264](https://github.com/JabRef/jabref/issues/1264): S with caron does not render correctly
+- LaTeX to Unicode converter now handles combining accents
 
 ### Removed